polishing

+++ fantastic-coffee-decaffeinated/service/database/remove-comment.go
@@ -1,11 +1,22 @@
 package database
 
 import (
+	"database/sql"
+	"errors"
 	"fmt"
 )
 
 // Delete a follow
 func (db *appdbimpl) RemoveComment(idcomment string) (string, error) {
+	var username string
+
+	rows := db.c.QueryRow("SELECT User FROM Comment WHERE Id_comment=?", idcomment).Scan(&username)
+
+	if errors.Is(rows, sql.ErrNoRows) {
+		// 404 comment not found
+		return "comment not found", rows
+	}
+
 	_, err := db.c.Exec("DELETE FROM Comment WHERE Id_comment = ?;", idcomment)
 
 	if err != nil {
