+++ fantastic-coffee-decaffeinated/service/database/unban-user.go
@@ -1,13 +1,22 @@
 package database
 
 import (
+	"database/sql"
+	"errors"
 	"fmt"
 )
 
 func (db *appdbimpl) UnbanUser(banner string, banned string) (string, error) {
+	var bannedUser string
+
+	rows := db.c.QueryRow("SELECT Banner FROM Ban WHERE Banned=?", banned).Scan(&bannedUser)
+
+	if errors.Is(rows, sql.ErrNoRows) {
+		// 404 like not found
+		return "Ban not found", rows
+	}
 
 	sqlStmt := fmt.Sprintf("DELETE FROM Ban WHERE Banner='%s' AND Banned='%s';", banner, banned)
-
 	_, err := db.c.Exec(sqlStmt)
 
 	if err != nil {
